+++ src/adapter/graphql/src/mutations/dataset_mut/adapters_mut/collection_mut.rs
@@ -12,20 +12,20 @@ use kamu::domain;
 use odf::metadata::OperationType as Op;
 
 use crate::prelude::*;
-use crate::queries::CollectionEntry;
+use crate::queries::{CollectionEntry, DatasetRequestState};
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
 #[derive(Debug)]
 pub struct CollectionMut<'a> {
-    dataset: &'a domain::ResolvedDataset,
+    state: &'a DatasetRequestState,
 }
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
 impl<'a> CollectionMut<'a> {
-    pub fn new(dataset: &'a domain::ResolvedDataset) -> Self {
-        Self { dataset }
+    pub fn new(state: &'a DatasetRequestState) -> Self {
+        Self { state }
     }
 
     // Push ingest the new record
@@ -39,6 +39,8 @@ impl<'a> CollectionMut<'a> {
     ) -> Result<CollectionUpdateResult> {
         use std::io::Write;
 
+        let dataset = self.state.resolved_dataset(ctx).await?;
+
         let push_ingest_use_case = from_catalog_n!(ctx, dyn domain::PushIngestDataUseCase);
 
         let mut ndjson = Vec::<u8>::new();
@@ -50,7 +52,7 @@ impl<'a> CollectionMut<'a> {
 
         let ingest_result = match push_ingest_use_case
             .execute(
-                self.dataset,
+                dataset,
                 kamu_core::DataSource::Buffer(bytes::Bytes::from_owner(ndjson)),
                 kamu_core::PushIngestDataUseCaseOptions {
                     source_name: None,
@@ -120,7 +122,7 @@ impl<'a> CollectionMut<'a> {
         // TODO: PERF: Filter paths that are relevant to operations
         let query_res = query_svc
             .get_data(
-                &self.dataset.get_handle().as_local_ref(),
+                &self.state.dataset_handle().as_local_ref(),
                 domain::GetDataOptions::default(),
             )
             .await