+++ src/infra/core/tests/tests/test_query_service_impl.rs
@@ -84,7 +84,7 @@ async fn create_test_dataset(
         .dataset
         .commit_event(
             MetadataFactory::set_data_schema()
-                .schema(&schema)
+                .schema_from_arrow(&schema)
                 .build()
                 .into(),
             odf::dataset::CommitOpts::default(),
@@ -124,6 +124,7 @@ async fn create_test_dataset(
                         start: start_offset,
                         end: end_offset,
                     }),
+                    new_linked_objects: None,
                     new_watermark: None,
                     new_source_state: None,
                 },
@@ -205,7 +206,7 @@ async fn test_dataset_parquet_schema(catalog: &dill::Catalog, tempdir: &TempDir)
 
     let query_svc = catalog.get_one::<dyn QueryService>().unwrap();
     let schema = query_svc
-        .get_schema_parquet_file(&dataset_ref)
+        .get_last_data_chunk_schema_parquet(&dataset_ref)
         .await
         .unwrap();
     assert!(schema.is_some());
@@ -236,43 +237,19 @@ async fn test_dataset_parquet_schema(catalog: &dill::Catalog, tempdir: &TempDir)
     );
 }
 
-async fn test_dataset_arrow_schema(catalog: &dill::Catalog, tempdir: &TempDir) {
+async fn test_dataset_schema(catalog: &dill::Catalog, tempdir: &TempDir) {
     let target = create_test_dataset(catalog, tempdir.path(), "foo").await;
     let dataset_ref = odf::DatasetRef::from(target.get_alias());
 
     let query_svc = catalog.get_one::<dyn QueryService>().unwrap();
-    let schema_ref = query_svc.get_schema(&dataset_ref).await.unwrap().unwrap();
+    let schema = query_svc.get_schema(&dataset_ref).await.unwrap().unwrap();
 
-    let mut buf = Vec::new();
-    odf::utils::schema::format::write_schema_arrow_json(&mut buf, schema_ref.as_ref()).unwrap();
-
-    let schema_content = String::from_utf8(buf).unwrap();
-    let data_schema_json =
-        serde_json::from_str::<serde_json::Value>(schema_content.as_str()).unwrap();
-
-    assert_eq!(
-        data_schema_json,
-        serde_json::json!({
-            "fields": [
-                {
-                    "name": "offset",
-                    "data_type": "UInt64",
-                    "nullable": false,
-                    "dict_id": 0,
-                    "dict_is_ordered": false,
-                    "metadata": {}
-                },
-                {
-                    "name": "blah",
-                    "data_type": "Utf8",
-                    "nullable": false,
-                    "dict_id": 0,
-                    "dict_is_ordered": false,
-                    "metadata": {}
-                }
-            ],
-            "metadata": {}
-        })
+    odf::utils::testing::assert_odf_schema_eq(
+        &schema,
+        &odf::schema::DataSchema::new(vec![
+            odf::schema::DataField::u64("offset"),
+            odf::schema::DataField::string("blah"),
+        ]),
     );
 }
 
@@ -307,9 +284,16 @@ async fn test_dataset_parquet_schema_local_fs() {
 
 #[test_group::group(engine, datafusion)]
 #[test_log::test(tokio::test)]
-async fn test_dataset_arrow_schema_local_fs() {
+async fn test_dataset_schema_local_fs() {
     let (tempdir, catalog) = prepare_schema_test_catalog();
-    test_dataset_arrow_schema(&catalog, &tempdir).await;
+    test_dataset_schema(&catalog, &tempdir).await;
+}
+
+#[test_group::group(containerized, engine, datafusion)]
+#[test_log::test(tokio::test)]
+async fn test_dataset_schema_s3() {
+    let (s3, catalog) = prepare_schema_test_s3_catalog().await;
+    test_dataset_schema(&catalog, &s3.tmp_dir).await;
 }
 
 #[test_group::group(containerized, engine, datafusion)]
@@ -317,13 +301,6 @@ async fn test_dataset_arrow_schema_local_fs() {
 async fn test_dataset_parquet_schema_s3() {
     let (s3, catalog) = prepare_schema_test_s3_catalog().await;
     test_dataset_parquet_schema(&catalog, &s3.tmp_dir).await;
-}
-
-#[test_group::group(containerized, engine, datafusion)]
-#[test_log::test(tokio::test)]
-async fn test_dataset_arrow_schema_s3() {
-    let (s3, catalog) = prepare_schema_test_s3_catalog().await;
-    test_dataset_arrow_schema(&catalog, &s3.tmp_dir).await;
 }
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////