+++ src/odf/metadata/src/dtos/mod.rs
@@ -10,7 +10,9 @@
 mod constants;
 mod dtos_extra;
 mod dtos_generated;
+mod extra_attributes_impl;
 mod operation_type;
 
 pub use dtos_generated::*;
+pub use extra_attributes_impl::*;
 pub use operation_type::*;