+++ src/odf/odf/src/lib.rs
@@ -47,6 +47,10 @@ pub use odf_metadata::{
 };
 pub use odf_storage::{GetBlockError, GetRefError};
 
+pub mod schema {
+    pub use odf_metadata::schema::*;
+}
+
 pub mod metadata {
     pub use odf_metadata::*;
 }